test/requirements.txt
azure-storage-blob
python-dotenv
openai
comtypes ; sys_platform == 'win32'
streamlit